requirements.txt
Flask
gunicorn
python-dotenv
langchain-community
langchain-groq
langchain-core
PyMuPDF
fitz